WIP Add semaphore coordinates for letters
Currently a-d has been tested, e-g is still WIP.

To get the same servo positions, angles on the left and right need to be supplementary. So 0 on the left would be 180 on the right, which is unintuitive. This could all change because we're planning to rework how the angles map to servos positions as mentioned in https://github.com/pdx-robotics/Arduino_KHR-1/issues/4