Optimize semaphore function for memory usage
KHR_1 can only semaphore short strings because all the animations for every char are appended to a large animation, which is then passed into animate().

In this commit, semaphore() calls animate() on every little animation, letting it go out of scope and deallocated afterwards.

The code compiles, but it's untested. Hopefully this will help KHR_1 support longer strings.