Updated code to match servo diagram.
Changed the defines in lib/Arduino_KHR1.h to match the servo diagram at the top and the servo reference.
However, due to the limitations of VarSpeedServo.h the head w of the KHR_1 was set to CH- 8 and the servos below the chest were commented out for the time being.